# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Risk & Disputes API
  version: 0.2.0
  description: Risk engine, fraud signals, dispute management, evidence, and representment
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3003
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/disputes/{id}/evidence:
    get:
      summary: List dispute evidence
      description: Returns paginated evidence for a dispute.
      x-platform:
        kind: resource
        entity: DisputeEvidence
        affectsLedger: false
      operationId: listDisputeEvidence
      tags:
        - DisputeEvidence
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of evidence.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DisputeEvidence"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dispute not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Add dispute evidence
      description: Adds a piece of evidence to an open dispute.
      x-platform:
        kind: resource
        entity: DisputeEvidence
        emitsEvents:
          - dispute.evidence.created
        affectsLedger: false
      operationId: addDisputeEvidence
      tags:
        - DisputeEvidence
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  enum:
                    - receipt
                    - shipping_proof
                    - customer_communication
                    - service_documentation
                    - policy_disclosure
                    - other
                file_name:
                  type: string
                  maxLength: 255
                file_size:
                  type: integer
                  minimum: 1
                description:
                  type: string
                  maxLength: 2000
      responses:
        "201":
          description: Evidence added.
          content:
            application/json:
              schema:
                type: object
                required:
                  - evidence
                properties:
                  evidence:
                    $ref: "#/components/schemas/DisputeEvidence"
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dispute not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Dispute is closed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/disputes/{id}/evidence/{evidenceId}:
    get:
      summary: Get evidence by ID
      description: Returns a single piece of evidence.
      x-platform:
        kind: resource
        entity: DisputeEvidence
        affectsLedger: false
      operationId: getDisputeEvidenceById
      tags:
        - DisputeEvidence
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: evidenceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evidence found.
          content:
            application/json:
              schema:
                type: object
                required:
                  - evidence
                properties:
                  evidence:
                    $ref: "#/components/schemas/DisputeEvidence"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dispute or evidence not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete evidence
      description: Deletes unsubmitted evidence from an open dispute.
      x-platform:
        kind: transition
        entity: DisputeEvidence
        emitsEvents:
          - dispute.evidence.deleted
        affectsLedger: false
        stateChanges:
          - evidence_removed
      operationId: deleteDisputeEvidence
      tags:
        - DisputeEvidence
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: evidenceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evidence deleted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deleted
                  - id
                properties:
                  deleted:
                    type: boolean
                  id:
                    type: string
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dispute or evidence not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Evidence already submitted or dispute closed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/disputes/{id}/representment:
    get:
      summary: Get representment
      description: Returns the representment for a dispute (paginated wrapper for consistency).
      x-platform:
        kind: resource
        entity: Representment
        affectsLedger: false
      operationId: getDisputeRepresentment
      tags:
        - Representment
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Representment list (0 or 1 items).
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Representment"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dispute not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create representment
      description: Creates a representment for a dispute. Only one per dispute.
      x-platform:
        kind: resource
        entity: Representment
        emitsEvents:
          - dispute.representment.created
        affectsLedger: false
      operationId: createRepresentment
      tags:
        - Representment
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                details:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Representment created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - representment
                properties:
                  representment:
                    $ref: "#/components/schemas/Representment"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dispute not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Representment already exists or dispute is closed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
    DisputeEvidence:
      description: A piece of evidence for a dispute.
      type: object
      properties:
        id:
          type: string
        dispute_id:
          type: string
        type:
          type: string
          enum:
            - receipt
            - shipping_proof
            - customer_communication
            - service_documentation
            - policy_disclosure
            - other
        file_name:
          type: string
          nullable: true
        file_size:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        submitted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Representment:
      description: A dispute representment.
      type: object
      properties:
        id:
          type: string
        dispute_id:
          type: string
        submitted_at:
          type: string
          format: date-time
          nullable: true
        outcome:
          type: string
          nullable: true
        details:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...